survey diff view: format size class
